# openapi.yaml
openapi: 3.0.1
info:
  title: SingleStore SQrL Plugin
  description: A plugin that allows the user to run SQL queries on a MySQL database (SingleStore) using ChatGPT.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /sql_query:
    post:
      operationId: executeSqlQuery
      summary: Execute a SQL query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sql_query:
                  type: string
                  description: The SQL query to execute.
                  required: true
                db_name:
                  type: string
                  description: The name of the database to execute the query on.
                  required: true
      responses:
        "200":
          description: OK
  /table_shape:
    get:
      operationId: getTableShape
      summary: Get the table names and column names. Necessary to determine the correct table/column names, so run this whenever you are unsure.
      responses:
        content:
              application/json:
                schema:
                  $ref: '#/components/schemas/getTableShapeResponse'
  /credentials:
    post:
      operationId: saveCredentials
      summary: Save database credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The database username.
                  required: false
                password:
                  type: string
                  description: The database password.
                  required: false
                endpoint_url:
                  type: string
                  description: The endpoint url of the database.
                  required: false

components:
  schemas:
    tableShapeResponse:
      type: object
      properties:
        todos:
          type: array
          items:
            type: string
          description: The list of column names, grouped by database name. 
    sql_query:
      type: object
      properties:
        sql_query:
          type: string
          description: The SQL query to execute.

    